requirements.txt
APScheduler==3.6.1
asn1crypto==1.0.0
astroid==2.3.1
autopep8==1.4.4
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
future==0.17.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.2
python-telegram-bot==12.1.1
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
tornado==6.0.3
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.6
wrapt==1.11.2
